test: finish organizing playwright suggestions tests into per feature modules
